\name{WISH_beta-package}
\alias{WISH_beta-package}
\alias{WISH_beta}
\docType{package}
\title{
\packageTitle{WISH_beta}
}
\description{
\packageDescription{WISH_beta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WISH_beta}
\packageIndices{WISH_beta}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{WISH_beta}

Maintainer: \packageMaintainer{WISH_beta}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
